requirements.txt
# Core LLM & Orchestration
langgraph>=0.2.0
langchain>=0.3.0
openai>=1.0.0
groq>=0.4.0
google-genai>=0.8.0  # Google Gemini
cohere>=5.0.0  # Cohere
ai21>=4.3.0  # AI21 Labs
# Alternative: azure-openai>=1.0.0

# Data Ingestion & API
fastapi>=0.104.0
jinja2>=3.1.2
websockets>=12.0  # For crypto data feed (Binance WebSocket)
websocket-client>=1.6.0  # For Zerodha WebSocket
uvicorn>=0.24.0
kiteconnect>=4.0.0

# Storage
redis>=5.0.0  # Includes async support (redis.asyncio)
pymongo>=4.6.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0
# Alternative: ta-lib (requires separate installation)

# Configuration & Validation
python-dotenv>=1.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
httpx>=0.25.0
aiohttp>=3.9.0
python-dateutil>=2.8.0
matplotlib>=3.7.0  # For backtest visualization
psutil>=5.9.0  # Process management (optional, used by start scripts)

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
feedparser>=6.0.10  # For RSS feed parsing
yfinance>=0.2.0  # For Yahoo Finance news and data
